v1.0.0 release
The generated SVGs and Lilypond source files will now be available
as an NPM package hosted through GitHub. This package will be used in
other LUTE applications such as the In C score progress viewer and the
in-development web app.

The SVGs are also pre-optimized using SVGO.

Part of me wants to rewrite the build script as a Node script but
replacing the file I/O was going to result in way more code than the
current build script with no real advantages, so I'm leaving it as-is
for now.